Fix Tmux layer macro
Previously, when we tried to do Shift-C, the macro actually ended up
sending something like this.

  KEY-DOWN - QMK: KC_LSFT
  KEY-DOWN - QMK: KC_LCTL
  KEY-DOWN - QMK: KC_B
  KEY-UP   - QMK: KC_B
  KEY-UP   - QMK: KC_LCTL
  KEY-DOWN - QMK: KC_C
  KEY-UP   - QMK: KC_C
  KEY-UP   - QMK: KC_LSFT

That is to say, it would technically send Ctrl-Shift-B, not Ctrl-B.

With this change, it sends this instead.

  KEY-DOWN - QMK: KC_LSFT
  KEY-DOWN - QMK: KC_LCTL
  KEY-UP   - QMK: KC_LSFT
  KEY-DOWN - QMK: KC_B
  KEY-UP   - QMK: KC_B
  KEY-UP   - QMK: KC_LCTL
  KEY-DOWN - QMK: KC_LSFT
  KEY-DOWN - QMK: KC_C
  KEY-UP   - QMK: KC_C
  KEY-UP   - QMK: KC_LSFT

Meaning it will deregister the Shift to only send Ctrl-B, then
re-register Shift so that the Shift-C comes through.